+++ climpred/options.py
@@ -26,7 +26,7 @@ _VALIDATORS = {
     in [True, False, "default"],
     "climpred_warnings": lambda choice: choice in [True, False, "default"],
     "bootstrap_resample_skill_func": lambda choice: choice
-    in ["loop", "empty_dim", "resample_before", "default"],
+    in ["loop", "exclude_resample_dim_from_dim", "resample_before", "default"],
     "resample_iterations_func": lambda choice: choice
     in ["default", "resample_iterations", "resample_iterations_idx"],
     "bootstrap_uninitialized_from_iterations_mean": lambda choice: choice
@@ -52,7 +52,7 @@ class set_options:
             If ``False`` use :py:func:`~climpred.reference.compute_persistence`.
             If ``True`` use
             :py:func:`~climpred.reference.compute_persistence_from_first_lead`.
-        ``warn_for_failed_PredictionEnsemble_xr_call`` : {``True``, ``False``}, default ``True``
+        ``warn_for_failed_PredictionEnsemble_xr_call`` : {``True``, ``False``}, default ``True``. # noqa: E501
             Raise ``UserWarning`` when ``PredictionEnsemble.xr_call``, e.g.
             ``.sel(lead=[1])`` fails on one of the datasets.
         ``warn_for_rename_to_climpred_dims`` : {``True``, ``False``}, default ``True``
@@ -64,28 +64,32 @@ class set_options:
             :py:class:`.PredictionEnsemble` is instantiated.
         ``climpred_warnings`` : {``True``, ``False``}, default ``True``
             Overwrites all options containing ``"*warn*"``.
-        ``bootstrap_resample_skill_func`` : {"loop", "empty_dim", "resample_before","default"} Decide which resampling method to use in
+        ``bootstrap_resample_skill_func`` : {"loop", "exclude_resample_dim_from_dim", "resample_before","default"}  # noqa: E501
+            Decide which resampling method to use in
             PredictionEnsemble.bootstrap(). ``default`` as in code.
 
             * ``loop`` calls :py:func:`climpred.bootstrap.resample_skill_loop` which
                 loops over iterations and calls ``verify`` every single time. Most
                 understandable and stable, but slow.
-            * ``empty_dim`` calls
-                :py:func:`climpred.bootstrap.resample_skill_empty_dim` which calls
-                ``verify(dim=dim_without_resample_dim)``, resamples over
+            * ``exclude_resample_dim_from_dim`` calls
+                :py:func:`climpred.bootstrap.resample_skill_exclude_resample_dim_from_dim`
+                which calls ``verify(dim=dim_without_resample_dim)``, resamples over
                 ``resample_dim`` and then takes a mean over ``resample_dim`` if in
                 ``dim``. Enables ``HindcastEnsemble.bootstrap(resample_dim="init", alignment="same_verifs")``.
                 Fast alternative for ``resample_dim="init"``.
-            * ``resample_before`` calls :py:func:`climpred.bootstrap.resample_skill_resample_before` which
+            * ``resample_before`` calls
+                :py:func:`climpred.bootstrap.resample_skill_resample_before` which
                 resamples ``iteration`` dimension and then calls ``verify`` vectorized.
                 Fast alternative for ``resample_dim="member"``.
 
-        ``resample_iterations_func``: {"default", 'resample_iterations',
-            'resample_iterations_idx'} Decide which resample_iterations function to use
-            from xskillscore. ``default`` as in code:
+        ``resample_iterations_func``: {``"default"``, ``"resample_iterations"``, ``"resample_iterations_idx"``}  # noqa: E501
+            Decide which resample_iterations function to use from xskillscore.
+            ``"default"`` as in code:
 
-            * :py:func:`xskillscore.resample_iterations_idx` creates one large chunk and consumes much memory and is not recommended for large files
-            * :py:func:`xskillscore.resample_iterations` create many tasks but is more stable
+            * :py:func:`xskillscore.resample_iterations_idx` creates one large chunk
+                and consumes much memory and is not recommended for large files.
+            * :py:func:`xskillscore.resample_iterations` create many tasks but is more
+                stable.
 
         ``bootstrap_uninitialized_from_iterations_mean``: {``True``, ``False``}
             Exchange ``uninitialized`` skill with the iteration mean ``uninitialized``.